st/mesa: use PRId64 for printing 64-bit ints
v2: use signed types/formats

+++ src/mesa/state_tracker/st_cb_bufferobjects.c
@@ -31,6 +31,8 @@
  */
 
 
+#include <inttypes.h>  /* for PRId64 macro */
+
 #include "main/imports.h"
 #include "main/mtypes.h"
 #include "main/arrayobj.h"
@@ -271,7 +273,8 @@ st_bufferobj_data(struct gl_context *ctx,
    pipe_resource_reference( &st_obj->buffer, NULL );
 
    if (ST_DEBUG & DEBUG_BUFFER) {
-      debug_printf("Create buffer size %td bind 0x%x\n", size, bind);
+      debug_printf("Create buffer size %" PRId64 " bind 0x%x\n",
+                   (int64_t) size, bind);
    }
 
    if (size != 0) {
